refactor(arel): converge the Hash analogue into rubyClassName
Move Dot#isHash's prototype-chain ancestor walk into ruby-class.ts as the
shared isHashAnalogue helper, so every visitor (ToSql/Mysql/PostgreSQL, not
just Dot) classifies a record derived from a plain record as a Hash — matching
Rails' object.class.ancestors walk (visitor.rb:36-41).

Guard the ctor-keyed dispatch path against a non-function constructor (an
inherited literal `constructor` key), which now reaches Visitor#visit directly
instead of being intercepted by Dot's pre-dispatch arm.

Deletes the isHash pre-dispatch residue from Dot#visit and Dot#isHash.
